pyproject.toml
[project]
name = "light-3d-localization"
version = "1.0.0"
description = "灯具3D定位系统 - 基于纯视觉方案的性能基线评估"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kagglehub",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "scipy>=1.11.0",
    "h5py>=3.9.0",
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "timm>=0.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "opencv-contrib-python>=4.8.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
light-3d-test = "test_environment:main"
light-3d-run = "run_all:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.uv]
dev-dependencies = []
